requirements.txt
numpy
langchain
langchain_groq
langchain_core
langchain_community
pymongo[srv]
fastapi
uvicorn
python-multipart
python-dotenv
passlib[bcrypt]